// lib/api/generated/models/CMSReleaseValidationOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CMSReleaseValidationOut = {
    release_id: string;
    valid: boolean;
    errors?: Array<Record<string, any>>;
    warnings?: Array<Record<string, any>>;
};